refactor: move ci_bump_command and ci_detect_command to versioning.py
