+++ docs/PATTERNS.md
@@ -143,6 +143,8 @@ def update_pyproject_version(path: Path, new_version: str) -> None: ...
 def update_init_version(path: Path, new_version: str) -> None: ...
 def create_tag(version: str, cwd: Path | None = None) -> None: ...
 def detect_bump_type(before_sha: str, after_sha: str, cwd: Path | None = None) -> str | None: ...
+def ci_bump_command(root: Path, before_sha: str, after_sha: str = "HEAD", dry_run: bool = False, no_tag: bool = False) -> None: ...
+def ci_detect_command(root: Path, before_sha: str, after_sha: str = "HEAD") -> None: ...
 ```
 
 **promise_verify.py:**
